refactor: :recycle: 规整CURD组件
