+++ src/typings/auto-imports.d.ts
@@ -325,14 +325,12 @@ declare module "vue" {
   interface GlobalComponents {}
   interface ComponentCustomProperties {
     readonly EffectScope: UnwrapRef<(typeof import("vue"))["EffectScope"]>;
-    readonly ElForm: UnwrapRef<(typeof import("element-plus/es"))["ElForm"]>;
     readonly ElMessage: UnwrapRef<
       (typeof import("element-plus/es"))["ElMessage"]
     >;
     readonly ElMessageBox: UnwrapRef<
       (typeof import("element-plus/es"))["ElMessageBox"]
     >;
-    readonly ElTree: UnwrapRef<(typeof import("element-plus/es"))["ElTree"]>;
     readonly acceptHMRUpdate: UnwrapRef<
       (typeof import("pinia"))["acceptHMRUpdate"]
     >;
@@ -1055,14 +1053,12 @@ declare module "@vue/runtime-core" {
   interface GlobalComponents {}
   interface ComponentCustomProperties {
     readonly EffectScope: UnwrapRef<(typeof import("vue"))["EffectScope"]>;
-    readonly ElForm: UnwrapRef<(typeof import("element-plus/es"))["ElForm"]>;
     readonly ElMessage: UnwrapRef<
       (typeof import("element-plus/es"))["ElMessage"]
     >;
     readonly ElMessageBox: UnwrapRef<
       (typeof import("element-plus/es"))["ElMessageBox"]
     >;
-    readonly ElTree: UnwrapRef<(typeof import("element-plus/es"))["ElTree"]>;
     readonly acceptHMRUpdate: UnwrapRef<
       (typeof import("pinia"))["acceptHMRUpdate"]
     >;
