+++ src/typings/components.d.ts
@@ -11,6 +11,7 @@ declare module "vue" {
     AppMain: (typeof import("./../layout/components/AppMain/index.vue"))["default"];
     BarChart: (typeof import("./../views/dashboard/components/BarChart.vue"))["default"];
     Breadcrumb: (typeof import("./../components/Breadcrumb/index.vue"))["default"];
+    CURD: (typeof import("./../components/CURD/index.vue"))["default"];
     DeptTree: (typeof import("./../views/system/user/components/dept-tree.vue"))["default"];
     Dictionary: (typeof import("./../components/Dictionary/index.vue"))["default"];
     DictItem: (typeof import("./../views/system/dict/components/dict-item.vue"))["default"];
@@ -53,16 +54,13 @@ declare module "vue" {
     ElSwitch: (typeof import("element-plus/es"))["ElSwitch"];
     ElTable: (typeof import("element-plus/es"))["ElTable"];
     ElTableColumn: (typeof import("element-plus/es"))["ElTableColumn"];
-    ElTabPane: (typeof import("element-plus/es"))["ElTabPane"];
-    ElTabs: (typeof import("element-plus/es"))["ElTabs"];
     ElTag: (typeof import("element-plus/es"))["ElTag"];
     ElText: (typeof import("element-plus/es"))["ElText"];
     ElTooltip: (typeof import("element-plus/es"))["ElTooltip"];
-    ElTree: (typeof import("element-plus/es"))["ElTree"];
     ElTreeSelect: (typeof import("element-plus/es"))["ElTreeSelect"];
     ElUpload: (typeof import("element-plus/es"))["ElUpload"];
     ElWatermark: (typeof import("element-plus/es"))["ElWatermark"];
-    Form: (typeof import("./../components/PageModal/Form.vue"))["default"];
+    Form: (typeof import("./../components/CURD/Form.vue"))["default"];
     FunnelChart: (typeof import("./../views/dashboard/components/FunnelChart.vue"))["default"];
     GithubCorner: (typeof import("./../components/GithubCorner/index.vue"))["default"];
     Hamburger: (typeof import("./../components/Hamburger/index.vue"))["default"];
@@ -70,26 +68,17 @@ declare module "vue" {
     IEpArrowDown: (typeof import("~icons/ep/arrow-down"))["default"];
     IEpArrowUp: (typeof import("~icons/ep/arrow-up"))["default"];
     IEpClose: (typeof import("~icons/ep/close"))["default"];
-    IEpDelete: (typeof import("~icons/ep/delete"))["default"];
     IEpDownload: (typeof import("~icons/ep/download"))["default"];
-    IEpEdit: (typeof import("~icons/ep/edit"))["default"];
-    IEpPlus: (typeof import("~icons/ep/plus"))["default"];
-    IEpPosition: (typeof import("~icons/ep/position"))["default"];
-    IEpQuestionFilled: (typeof import("~icons/ep/question-filled"))["default"];
-    IEpRefresh: (typeof import("~icons/ep/refresh"))["default"];
-    IEpRefreshLeft: (typeof import("~icons/ep/refresh-left"))["default"];
-    IEpSearch: (typeof import("~icons/ep/search"))["default"];
-    IEpTop: (typeof import("~icons/ep/top"))["default"];
-    IEpUploadFilled: (typeof import("~icons/ep/upload-filled"))["default"];
     LangSelect: (typeof import("./../components/LangSelect/index.vue"))["default"];
     LayoutSelect: (typeof import("./../layout/components/Settings/components/LayoutSelect.vue"))["default"];
     MultiUpload: (typeof import("./../components/Upload/MultiUpload.vue"))["default"];
     NavBar: (typeof import("./../layout/components/NavBar/index.vue"))["default"];
     NavbarLeft: (typeof import("./../layout/components/NavBar/components/NavbarLeft.vue"))["default"];
     NavbarRight: (typeof import("./../layout/components/NavBar/components/NavbarRight.vue"))["default"];
-    PageContent: (typeof import("./../components/PageContent/index.vue"))["default"];
-    PageModal: (typeof import("./../components/PageModal/index.vue"))["default"];
-    PageSearch: (typeof import("./../components/PageSearch/index.vue"))["default"];
+    PageContent: (typeof import("./../components/CURD/PageContent.vue"))["default"];
+    PageForm: (typeof import("./../components/CURD/PageForm.vue"))["default"];
+    PageModal: (typeof import("./../components/CURD/PageModal.vue"))["default"];
+    PageSearch: (typeof import("./../components/CURD/PageSearch.vue"))["default"];
     Pagination: (typeof import("./../components/Pagination/index.vue"))["default"];
     PieChart: (typeof import("./../views/dashboard/components/PieChart.vue"))["default"];
     RadarChart: (typeof import("./../views/dashboard/components/RadarChart.vue"))["default"];
